packages/eventbrite_service/pyproject.toml
[tool.poetry]
name = "eventbrite-service"
version = "0.1.0"
description = ""
authors = ["sajeremy <9jsantiago4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
ruff = "^0.8.0"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
loguru = "^0.7.2"
selenium = "^4.27.1"

[tool.ruff]
src=["packages/eventbrite_service"]
line-length=120
select=["E","F","I","W"]
ignore=["E501"]
fix=true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
